requirements.txt
# Nexus — Multi-Agent Vendor Verification System
# All dependencies are free and open-source

# === Agent Framework ===
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-core>=0.3.0

# === Backend API ===
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
httpx>=0.27.0
python-multipart>=0.0.12

# === Data Models ===
pydantic>=2.0.0
pydantic-settings>=2.0.0

# === Document Processing ===
PyMuPDF>=1.24.0
reportlab>=4.1.0

# === Async Support ===
asyncio-extras>=1.3.0

# === Utilities ===
uuid6>=2024.1.12
